Make theme store a model
