+++ packages/devtools-ui/src/context.ts
@@ -1,10 +1,10 @@
-import { signal, computed, effect } from "@preact/signals";
 import type { DevToolsAdapter } from "@preact/signals-devtools-adapter";
 import { ConnectionModel } from "./models/ConnectionModel";
 import { SettingsModel } from "./models/SettingsModel";
+import { ThemeModel } from "./models/ThemeModel";
 import { UpdatesModel } from "./models/UpdatesModel";
 
-export { ConnectionModel, SettingsModel, UpdatesModel };
+export { ConnectionModel, SettingsModel, ThemeModel, UpdatesModel };
 export type {
 	SignalUpdate,
 	Divider,
@@ -12,15 +12,14 @@ export type {
 	UpdateTreeNodeSingle,
 	UpdateTreeNodeGroup,
 } from "./models/UpdatesModel";
-
-export type ThemeMode = "auto" | "light" | "dark";
+export type { ThemeMode } from "./models/ThemeModel";
 
 export interface DevToolsContext {
 	adapter: DevToolsAdapter;
 	connectionStore: InstanceType<typeof ConnectionModel>;
 	updatesStore: InstanceType<typeof UpdatesModel>;
 	settingsStore: InstanceType<typeof SettingsModel>;
-	themeStore: ReturnType<typeof createThemeStore>;
+	themeStore: InstanceType<typeof ThemeModel>;
 }
 
 let currentContext: DevToolsContext | null = null;
@@ -34,78 +33,11 @@ export function getContext(): DevToolsContext {
 	return currentContext;
 }
 
-const THEME_STORAGE_KEY = "signals-devtools-theme";
-
-export function createThemeStore() {
-	const stored = (() => {
-		try {
-			const val = localStorage.getItem(THEME_STORAGE_KEY);
-			if (val === "light" || val === "dark" || val === "auto") return val;
-		} catch {
-			// localStorage unavailable
-		}
-		return "auto" as ThemeMode;
-	})();
-
-	const theme = signal<ThemeMode>(stored);
-
-	const mediaQuery =
-		typeof window !== "undefined"
-			? window.matchMedia("(prefers-color-scheme: dark)")
-			: null;
-	const systemIsDark = signal(mediaQuery?.matches ?? false);
-
-	if (mediaQuery) {
-		const handler = (e: MediaQueryListEvent) => {
-			systemIsDark.value = e.matches;
-		};
-		mediaQuery.addEventListener("change", handler);
-	}
-
-	const resolvedTheme = computed<"light" | "dark">(() =>
-		theme.value === "auto"
-			? systemIsDark.value
-				? "dark"
-				: "light"
-			: theme.value
-	);
-
-	// Apply data-theme attribute to the devtools container
-	effect(() => {
-		const resolved = resolvedTheme.value;
-		const el = document.querySelector(".signals-devtools");
-		if (el instanceof HTMLElement) {
-			el.dataset.theme = resolved;
-		}
-	});
-
-	const toggleTheme = () => {
-		const order: ThemeMode[] = ["auto", "light", "dark"];
-		const idx = order.indexOf(theme.value);
-		theme.value = order[(idx + 1) % order.length];
-		try {
-			localStorage.setItem(THEME_STORAGE_KEY, theme.value);
-		} catch {
-			// localStorage unavailable
-		}
-	};
-
-	return {
-		get theme() {
-			return theme.value;
-		},
-		get resolvedTheme() {
-			return resolvedTheme.value;
-		},
-		toggleTheme,
-	};
-}
-
 export function initDevTools(adapter: DevToolsAdapter): DevToolsContext {
 	const settingsStore = new SettingsModel(adapter);
 	const updatesStore = new UpdatesModel(adapter, settingsStore);
 	const connectionStore = new ConnectionModel(adapter);
-	const themeStore = createThemeStore();
+	const themeStore = new ThemeModel();
 
 	currentContext = {
 		adapter,
@@ -123,6 +55,7 @@ export function destroyDevTools(): void {
 		currentContext.connectionStore[Symbol.dispose]();
 		currentContext.updatesStore[Symbol.dispose]();
 		currentContext.settingsStore[Symbol.dispose]();
+		currentContext.themeStore[Symbol.dispose]();
 		currentContext.adapter.disconnect();
 		currentContext = null;
 	}
