+++ packages/devtools-ui/src/models/index.ts
@@ -1,5 +1,6 @@
 export { ConnectionModel } from "./ConnectionModel";
 export { SettingsModel } from "./SettingsModel";
+export { ThemeModel, type ThemeMode } from "./ThemeModel";
 export {
 	UpdatesModel,
 	type SignalUpdate,
